adding adapters

+++ src/main/java/fr/aresrpg/tofumanchou/domain/data/entity/player/Perso.java
@@ -141,7 +141,7 @@ public interface Perso extends Player {
 
 	void moveItem(MovedItem items);
 
-	void moveKama(int amount);
+	void moveKama(long amount);
 
 	void destroyItem(long uid, int amount);
